(MODULES-9578) Create ssh_authorized_key in root path
Previously, when the `target` property was set, the ssh_authorized_key
resource could not create directories/files within root-owned
paths. This behavior is due to the module switching context to the
user, then attempting to create the directory/file as the specified
user, ultimately failing because of insufficient permissions.

This commit removes the context change logic, and creates the
authorized_key file as `root`, then executes `FileUtils.chown` to make
it owned by the target user.

This commit also removes the functionality that created parent
directories, since it should not be in the scope of this module.